connected models to college_readiness_initiative db
